requirements.txt
urllib3==1.24.1
beautifulsoup4==4.6.3
Pillow==5.3.0
pytesseract==0.2.5